Refactor structure:
- create folder structure
- create more interesting page
