+++ .gitignore
@@ -1,5 +1,22 @@
+# NodeJS dev assets
 node_modules
 build
+coverage
+
+# Database
 db
+
+# For when we copy from/to some code
 *.tar.gz
 *.tgz
+
+# misc
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
