--- fe/src/App.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const App = () => {
-  return (
-    <>
-      <span>Hello, World!</span>
-    </>
-  );
-}
