+++ fe/src/config/constants.ts
@@ -1,0 +1,2 @@
+export const TITLE = "React Application with ExpressJS and MariaDB back-end";
+export const RELEASE_YEAR = 2021;
