+++ fe/src/pages/Home/index.css
@@ -1,0 +1,26 @@
+header {
+  flex: 2;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+header span {
+  font-size: x-large;
+}
+
+main {
+  flex: 20;
+  display: flex;
+  width: 80%;
+  margin: 0 auto;
+  padding-left: 1em;
+}
+
+main > span {
+  margin-top: 1em;
+}
+
+footer {
+  flex: 1;
+}
