+++ fe/src/pages/Home/index.test.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
-import {App} from './App';
+import { App } from '.';
+import { TITLE } from '../../config/constants';
 
 test('renders H1 text', () => {
   render(<App />);
-  const textElement = screen.getByText(/hello, world/i);
+  const textElement = screen.getByText(TITLE);
   expect(textElement).toBeInTheDocument();
 });
